Flyttade collision från packetparser till world samt fixade diverse irriterande warnings.

+++ client/src/pointerviewstate.cpp
@@ -19,7 +19,7 @@ void PointerViewState::windowEvent(sf::Event event, Engine & engine)
 
 std::shared_ptr<PlayerViewState> PointerViewState::update(Engine & engine)
 {
-	sf::Time walk_wait = sf::seconds(0.2);
+	sf::Time walk_wait = sf::seconds((float)0.2);
 	bool waited = walk_timer.getElapsedTime() > walk_wait;
 	if ((
 		key::pressed(sf::Keyboard::Left)
